% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lut_iris_atome_indcvi_za}
\alias{lut_iris_atome_indcvi_za}
\title{IRIS Lookup Table for France, millesime 2020 restricted to zone A: reg_code == "11"}
\format{
Data frame with 3 columns and 5,238 rows
\describe{
\item{iris_code}{Code IRIS du lieu de résidence, 9 character}
\item{atome_indcvi}{atome_indcvi du lieu de résidence s'il est défini, cv_code sinon, 9 ou 5 character}
\item{atome_indcvi_type}{Type de l'atome_indcvi du lieu de résidence s'il est défini, cv_code sinon, 9 ou 5 character}
}
}
\usage{
lut_iris_atome_indcvi_za
}
\description{
Look Up Table (lut) between iris_code and atome_indcvi, obtained by comparison of
administrative dataset from INSEE & RP INDCVI detail file 2020 for reg_code == "11".
}
\section{Versioning}{

This dataset is versioned using semantic versioning. Check attributes:
\code{attr(lut_iris_atome_indcvi_za, "version")}
Current version: 2020-v1.0
}

\examples{
  lut_iris_atome_indcvi_za
}
\keyword{datasets}
