% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapuce.data.2020-package.R
\docType{package}
\name{mapuce.data.2020-package}
\alias{mapuce.data.2020}
\alias{mapuce.data.2020-package}
\title{mapuce.data.2020: Data Package for Mapuce Calculations}
\description{
This package provides models and data needed for Mapuce calculations.
}
\author{
\strong{Maintainer}: Alexis Bourgeois \email{acidybee@proton.me} (\href{https://orcid.org/0009-0009-8513-1715}{ORCID})

}
\keyword{internal}
